Use classic script tags instead of next/script
scripts from next/script are not loading properly
