+++ pages/_document.tsx
@@ -47,6 +47,48 @@ export default class MyDocument extends Document {
             href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;600&display=swap"
             rel="stylesheet"
           />
+          <script
+            type="text/javascript"
+            id="hs-script-loader"
+            async
+            defer
+            src="//js.hs-scripts.com/7707659.js"
+          ></script>
+          <script
+            async
+            src="https://www.googletagmanager.com/gtag/js?id=UA-2664098-14"
+          ></script>
+          <script
+            dangerouslySetInnerHTML={{
+              __html: `
+              window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              gtag('config', 'UA-2664098-14');
+          `,
+            }}
+            defer={true}
+          ></script>
+          <script
+            dangerouslySetInnerHTML={{
+              __html: `!function(f,b,e,v,n,t,s)
+{if(f.fbq)return;n=f.fbq=function(){n.callMethod?
+n.callMethod.apply(n,arguments):n.queue.push(arguments)};
+if(!f._fbq)f._fbq=n;n.push=n;n.loaded=!0;n.version='2.0';
+n.queue=[];t=b.createElement(e);t.async=!0;
+t.src=v;s=b.getElementsByTagName(e)[0];
+s.parentNode.insertBefore(t,s)}(window,document,'script',
+'https://connect.facebook.net/en_US/fbevents.js');
+ fbq('init', '779038593488772'); 
+fbq('track', 'PageView');`,
+            }}
+            defer={true}
+          />
+          <script
+            dangerouslySetInnerHTML={{
+              __html: `(function(){ window.ldfdr = window.ldfdr || {}; (function(d, s, ss, fs){ fs = d.getElementsByTagName(s)[0]; function ce(src){ var cs = d.createElement(s); cs.src = src; setTimeout(function(){fs.parentNode.insertBefore(cs,fs)}, 1); } ce(ss); })(document, 'script', 'https://sc.lfeeder.com/lftracker_v1_bElvO73Akn6aZMqj.js'); })(); `,
+            }}
+          />
         </Head>
         <body>
           <Main />
@@ -56,8 +98,7 @@ export default class MyDocument extends Document {
             <img
               height="1"
               width="1"
-              src="https://www.facebook.com/tr?id=779038593488772&ev=PageView
-&noscript=1"
+              src="https://www.facebook.com/tr?id=779038593488772&ev=PageView&noscript=1"
             />
           </noscript>
         </body>
